@@ -3,7 +3,7 @@ require 'on_the_spot/on_the_spot_helpers'
 
 require 'action_controller'
 
-describe "OnTheSpot" do
+RSpec.describe "OnTheSpot" do
   describe "Helpers" do
     before(:each) do
 
@@ -12,121 +12,162 @@ describe "OnTheSpot" do
       end
 
       @tester = TestClass.new
-      @test_array_nr  = [[1,"abc"], [2, "def"], [3, "ghi"]]
+      @test_array_nr  = [[1,"abc"], [2, "def"], [3, "ghi"], [4, "Freddy's Nightmare"]]
       @test_array_str = [["key", "value"], ["key2", "value2"]]
     end
 
     context "lookup values from array" do
       it "should find value abc for key 1" do
-        @tester.lookup_display_value(@test_array_nr, 1).should == 'abc'
+        expect(@tester.lookup_display_value(@test_array_nr, 1)).to eq('abc')
       end
 
       it "should find value <value2> for key key2" do
-        @tester.lookup_display_value(@test_array_str, 'key2').should == 'value2'
+        expect(@tester.lookup_display_value(@test_array_str, 'key2')).to eq('value2')
       end
 
       it "should return an empty string if key does not exist" do
-        @tester.lookup_display_value(@test_array_str, 'key1').should == ''
+        expect(@tester.lookup_display_value(@test_array_str, 'key1')).to eq('')
+      end
+
+      it "can handle single quotes normally" do
+        expect(@tester.lookup_display_value(@test_array_nr, 4)).to eq("Freddy's Nightmare")
       end
 
     end
 
     context "convert array to json" do
       it "should convert correctly" do
-        @tester.convert_array_to_json(@test_array_nr, 1).should == "{ '1':'abc', '2':'def', '3':'ghi', 'selected':'1'}"
+        expect(@tester.convert_array_to_json(@test_array_nr, 1)).to eq("{\"1\":\"abc\",\"2\":\"def\",\"3\":\"ghi\",\"4\":\"Freddy's Nightmare\",\"selected\":1}")
+      end
+
+      it "convert an array containing an item with single quotes to valid JSON" do
+        test_array_with_single_quote = [[1, "tree"], [2, "bike"], [3, "John's hat"]]
+        json_str = @tester.convert_array_to_json(test_array_with_single_quote, 1)
+        expect(json_str).to  eq("{\"1\":\"tree\",\"2\":\"bike\",\"3\":\"John's hat\",\"selected\":1}")
       end
     end
 
     context "the helpers" do
       before(:each) do
-        @dummy = mock()
-        @dummy.stub!(:content).and_return('test')
-        @dummy.stub!(:id).and_return('123')
-        @tester.should_receive(:t).with('on_the_spot.ok').and_return("ok")
-        @tester.should_receive(:t).with('on_the_spot.cancel').and_return("cancel")
-        @tester.should_receive(:t).with('on_the_spot.tooltip').and_return("tooltip")
+        @dummy = double()
+        allow(@dummy).to receive(:content).and_return('test')
+        allow(@dummy).to receive(:id).and_return('123')
+        expect(@tester).to receive(:t).with('on_the_spot.ok').and_return("ok")
+        expect(@tester).to receive(:t).with('on_the_spot.cancel').and_return("cancel")
+        expect(@tester).to receive(:t).with('on_the_spot.tooltip').and_return("tooltip")
+        expect(@tester).to receive(:t).with('on_the_spot.form_css').and_return(nil)
+        expect(@tester).to receive(:t).with('on_the_spot.input_css').and_return(nil)
       end
 
       context "with standard route" do
         before(:each) do
-          @tester.should_receive(:url_for).with({:action => 'update_attribute_on_the_spot'}).and_return('/bla')
+          expect(@tester).to receive(:url_for).with({:action => 'update_attribute_on_the_spot'}).and_return('/bla')
         end
 
         it "makes the correct html for an edit-field" do
           @result = @tester.on_the_spot_edit @dummy, :content
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\">test</span>")
         end
 
         it "makes the correct html for an edit-field with text selected on click" do
           @result = @tester.on_the_spot_edit @dummy, :content, :selected => true
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-ok=\"ok\" data-selected=\"true\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-selected=\"true\">test</span>")
         end
 
         it "makes the correct html for an edit-field and overrule display-text" do
           @result = @tester.on_the_spot_edit @dummy, :content, :display_text => 'jediknight'
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">jediknight</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\">jediknight</span>")
         end
+
+        it "makes the correct html for an edit-field and use the display-method as string" do
+          expect(@dummy).to receive(:changed_content).and_return("test-changed")
+          expect(@tester).to receive(:url_for).with({:action => 'get_attribute_on_the_spot'}).and_return('/bla-again')
+
+          @result = @tester.on_the_spot_edit @dummy, :content, :display_method => 'changed_content'
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-display-method=\"changed_content\" data-loadurl=\"/bla-again\">test-changed</span>")
+        end
+
+        it "makes the correct html for an edit-field and use the display-method (as symbol)" do
+          allow(@dummy).to receive(:changed_content).and_return("test-changed")
+          allow(@tester).to receive(:url_for).with({:action => 'get_attribute_on_the_spot'}).and_return('/bla-again')
+
+          @result = @tester.on_the_spot_edit @dummy, :content, :display_method => :changed_content
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-display-method=\"changed_content\" data-loadurl=\"/bla-again\">test-changed</span>")
+        end
+
 
         it "makes the correct html for a text-area" do
           @result = @tester.on_the_spot_edit @dummy, :content, :type => :textarea
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-columns=\"40\" data-edittype=\"textarea\" data-ok=\"ok\" data-rows=\"5\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-edittype=\"textarea\" data-rows=\"5\" data-columns=\"40\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\">test</span>")
         end
 
         it "makes the correct html for a text-area with text selected on click" do
           @result = @tester.on_the_spot_edit @dummy, :content, :type => :textarea, :selected => true
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-columns=\"40\" data-edittype=\"textarea\" data-ok=\"ok\" data-rows=\"5\" data-selected=\"true\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-edittype=\"textarea\" data-rows=\"5\" data-columns=\"40\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-selected=\"true\">test</span>")
         end
 
         it "makes the correct html for a select-box" do
           @result = @tester.on_the_spot_edit @dummy, :content, :type => :select, :data => [['test', 'This a test'], ['prod', 'Pure Production'], ['QA', 'Quality Assurance']]
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-edittype=\"select\" data-ok=\"ok\" data-select=\"{ 'test':'This a test', 'prod':'Pure Production', 'QA':'Quality Assurance', 'selected':'test'}\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">This a test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-edittype=\"select\" data-select=\"{&quot;test&quot;:&quot;This a test&quot;,&quot;prod&quot;:&quot;Pure Production&quot;,&quot;QA&quot;:&quot;Quality Assurance&quot;,&quot;selected&quot;:&quot;test&quot;}\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\">This a test</span>")
         end
 
         it "makes the correct html for a checkbox" do
-          @dummy.stub!(:content).and_return(true)
+          expect(@dummy).to receive(:content).and_return(true)
           @result = @tester.on_the_spot_edit @dummy, :content, :type => :checkbox
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-edittype=\"checkbox\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">true</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-edittype=\"checkbox\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\">true</span>")
         end
 
         it "makes the correct html for an edit-field with a callback" do
           @result = @tester.on_the_spot_edit @dummy, :content, :callback => 'testCallback(value, settings);'
-          @result.should == "<span class=\"on_the_spot_editing\" data-callback=\"testCallback(value, settings);\" data-cancel=\"cancel\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-callback=\"testCallback(value, settings);\">test</span>")
         end
 
         it "makes the correct html for an edit-field with an onblur action" do
           @result = @tester.on_the_spot_edit @dummy, :content, :onblur => 'submit'
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-ok=\"ok\" data-onblur=\"submit\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-onblur=\"submit\">test</span>")
         end
 
         it "makes the correct html for an edit-field with a loadurl" do
           @result = @tester.on_the_spot_edit @dummy, :content, :loadurl => '/load/data'
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-loadurl=\"/load/data\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-loadurl=\"/load/data\">test</span>")
         end
 
         context "a select-box with a loadurl" do
           it "makea the correct html (and not look up the value)" do
             @result = @tester.on_the_spot_edit @dummy, :content, :type => :select, :loadurl => '/load/data'
-            @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-edittype=\"select\" data-loadurl=\"/load/data\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+            expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-edittype=\"select\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-loadurl=\"/load/data\">test</span>")
           end
 
-          it "usea the display-text preferrably" do
+          it "use the display-text preferrably" do
             @result = @tester.on_the_spot_edit @dummy, :content, :type => :select, :loadurl => '/load/data', :display_text => 'ninja'
-            @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-edittype=\"select\" data-loadurl=\"/load/data\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">ninja</span>"
+            expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-edittype=\"select\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\" data-loadurl=\"/load/data\">ninja</span>")
           end
         end
       end
 
       context "with explicit route" do
         before(:each) do
-          @tester.should_receive(:url_for).with({:action => 'update_it_otherwise'}).and_return('/bla')
+          allow(@tester).to receive(:url_for).with({:action => 'update_it_otherwise'}).and_return('/bla')
         end
 
         it "make the correct html for an edit-field" do
           @result = @tester.on_the_spot_edit @dummy, :content, :url => {:action => 'update_it_otherwise' }
-          @result.should == "<span class=\"on_the_spot_editing\" data-cancel=\"cancel\" data-ok=\"ok\" data-tooltip=\"tooltip\" data-url=\"/bla\" id=\"r_spec/mocks/mock__content__123\">test</span>"
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\">test</span>")
+        end
+      end
+
+      context "with raw parameter" do
+        before(:each) do
+          allow(@tester).to receive(:url_for).with({:action => 'update_attribute_on_the_spot', :raw => true}).and_return('/bla?raw=true')
         end
 
+        it "supports raw html" do
+          allow(@dummy).to receive(:content).and_return('<b>test</b>')
+          @result = @tester.on_the_spot_edit @dummy, :content, :raw => true
+          expect(@result).to eq("<span id=\"r_spec/mocks/double__content__123\" class=\"on_the_spot_editing\" data-url=\"/bla?raw=true\" data-ok=\"ok\" data-cancel=\"cancel\" data-tooltip=\"tooltip\"><b>test</b></span>")
+        end
       end
     end
   end
 end
+
